migrations functioning, tested

+++ backend/db/migrations/20230425070939-create-store-item.js
@@ -14,7 +14,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.DOUBLE(5, 2),
+        type: Sequelize.DECIMAL(5, 2),
         allowNull: false
       },
       description: {
